now the C++ bridge uses static type checking

+++ bridge/cpp/bxx/runtime.typechecker.hpp
@@ -28,6 +28,7 @@ namespace bxx {
 //
 //  Default to deny
 //
+/*
 template <size_t Opcode, typename Out, typename In1, typename In2>
 inline
 void Runtime::typecheck(void)
@@ -72,8 +73,8 @@ void Runtime::typecheck(void)
 
     throw std::runtime_error(ss.str());
 }
+*/
 
-/*
 template <size_t Opcode>
 struct dependent_false { enum { value = false }; };
 
@@ -94,7 +95,6 @@ void Runtime::typecheck(void)
 {
     static_assert(dependent_false<Opcode>::value, "ArrayOperation has unsupported type-signature.");
 }
-*/
 
 //
 //  Allowed types.
@@ -563,7 +563,7 @@ void Runtime::typecheck<BH_BITWISE_OR_REDUCE, bool, bool, int64_t>(void) { }
 
 template <>
 inline
-void Runtime::typecheck<BH_BITWISE_OR_REDUCE, int16_t, int16_t, int16_t>(void) { }
+void Runtime::typecheck<BH_BITWISE_OR_REDUCE, int16_t, int16_t, int64_t>(void) { }
 
 
 template <>
